fix: handle padded lines
For performance reasons, GStreamer pads lines to be multiples of 8 bytes
(this is actually very common). Even though Rock's datatype does
account for this with the row_size attribute, it is not consistent
and I am guessing that a lot of code would fail if we had padded data.

Convert to/from padded lines when necessary.