fix: handle asynchronicity in the start of the negotiation process
The current code was sending a message or two at the start of the
components to "announce" their presence. This is very limited, and
definitely does not cover all the possible cases. For instance:

- receiver connects to signalling and sends a request message
- sender is started without a remote_peer_id

This commit changes the scheme to handle negotiation timeouts. This
makes components essentially send offer and offer requests periodically,
until someone responds to them on the other side. Tests are written
to cover all the possible orderings.

I have run this code for an hour to reasonably make sure there are no
remaining race conditions in the signalling process.